wrap up attendee
